Progress bar updates as needed and shows TX and %.
The progress bar now shows the transfer rate and percentage complete
instead of the source and destination.

The updating of the progress bar every loop was killing the CPU performance
and causing the Xorg process to go crazy. By only updating the progress bar
if the percent complete value changes stop this from happening.